cmake fixes

+++ src/mir_printer.c
@@ -53,7 +53,7 @@ print_instr_head(MirInstr *instr, FILE *stream, const char *name)
             instr->ref_count);
   }
 #else
-  fprintf(stream, "  %%%-3u", instr->id);
+  fprintf(stream, "  %%%-3llu", (unsigned long long)instr->id);
 #endif
   print_type(instr->const_value.type, true, stream, true);
   fprintf(stream, " %s ", name);
@@ -222,7 +222,7 @@ print_comptime_value_or_id(MirInstr *instr, FILE *stream)
   }
 
   if (!instr->comptime || !instr->analyzed) {
-    fprintf(stream, "%%%u", instr->id);
+    fprintf(stream, "%%%llu", (unsigned long long)instr->id);
     return;
   }
 
@@ -332,14 +332,14 @@ print_instr_type_fn(MirInstrTypeFn *type_fn, FILE *stream)
     MirInstr *tmp;
     barray_foreach(type_fn->arg_types, tmp)
     {
-      fprintf(stream, "%%%llu", tmp->id);
+      fprintf(stream, "%%%llu", (unsigned long long)tmp->id);
       if (i + 1 < bo_array_size(type_fn->arg_types)) fprintf(stream, ", ");
     }
   }
 
   fprintf(stream, ")");
 
-  if (type_fn->ret_type) fprintf(stream, " %%%llu", type_fn->ret_type->id);
+  if (type_fn->ret_type) fprintf(stream, " %%%llu", (unsigned long long)type_fn->ret_type->id);
 }
 
 void
@@ -363,28 +363,29 @@ void
 print_instr_type_ptr(MirInstrTypePtr *type_ptr, FILE *stream)
 {
   print_instr_head(&type_ptr->base, stream, "const");
-  fprintf(stream, "*%%%u", type_ptr->type->id);
+  fprintf(stream, "*%%%llu", (unsigned long long)type_ptr->type->id);
 }
 
 void
 print_instr_type_array(MirInstrTypeArray *type_array, FILE *stream)
 {
   print_instr_head(&type_array->base, stream, "const");
-  fprintf(stream, "[%%%u]%%%u", type_array->len->id, type_array->elem_type->id);
+  fprintf(stream, "[%%%llu]%%%llu", (unsigned long long)type_array->len->id,
+          (unsigned long long)type_array->elem_type->id);
 }
 
 void
 print_instr_type_slice(MirInstrTypeSlice *type_slice, FILE *stream)
 {
   print_instr_head(&type_slice->base, stream, "const");
-  fprintf(stream, "[]%%%u", type_slice->elem_type->id);
+  fprintf(stream, "[]%%%llu", (unsigned long long)type_slice->elem_type->id);
 }
 
 void
 print_instr_type_vargs(MirInstrTypeVArgs *type_vargs, FILE *stream)
 {
   print_instr_head(&type_vargs->base, stream, "const");
-  fprintf(stream, "...%%%u", type_vargs->elem_type->id);
+  fprintf(stream, "...%%%llu", (unsigned long long)type_vargs->elem_type->id);
 }
 
 void
@@ -428,7 +429,7 @@ print_instr_cast(MirInstrCast *cast, FILE *stream)
     bl_unimplemented;
   }
 
-  fprintf(stream, "%%%u", cast->next->id);
+  fprintf(stream, "%%%llu", (unsigned long long)cast->next->id);
 }
 
 void
@@ -501,7 +502,7 @@ void
 print_instr_elem_ptr(MirInstrElemPtr *elem_ptr, FILE *stream)
 {
   print_instr_head(&elem_ptr->base, stream, "elemptr");
-  fprintf(stream, "%%%u[", elem_ptr->arr_ptr->id);
+  fprintf(stream, "%%%llu[", (unsigned long long)elem_ptr->arr_ptr->id);
   print_comptime_value_or_id(elem_ptr->index, stream);
   fprintf(stream, "]");
 }
@@ -550,8 +551,9 @@ print_instr_cond_br(MirInstrCondBr *cond_br, FILE *stream)
 {
   print_instr_head(&cond_br->base, stream, "br");
   print_comptime_value_or_id(cond_br->cond, stream);
-  fprintf(stream, " ? %%%s_%u : %%%s_%u", cond_br->then_block->name, cond_br->then_block->base.id,
-          cond_br->else_block->name, cond_br->else_block->base.id);
+  fprintf(stream, " ? %%%s_%llu : %%%s_%llu", cond_br->then_block->name,
+          (unsigned long long)cond_br->then_block->base.id, cond_br->else_block->name,
+          (unsigned long long)cond_br->else_block->base.id);
 }
 
 void
@@ -571,7 +573,7 @@ void
 print_instr_br(MirInstrBr *br, FILE *stream)
 {
   print_instr_head(&br->base, stream, "br");
-  fprintf(stream, "%%%s_%d", br->then_block->name, br->then_block->base.id);
+  fprintf(stream, "%%%s_%llu", br->then_block->name, (unsigned long long)br->then_block->base.id);
 }
 
 void
@@ -585,7 +587,7 @@ void
 print_instr_addrof(MirInstrAddrOf *addrof, FILE *stream)
 {
   print_instr_head(&addrof->base, stream, "addrof");
-  fprintf(stream, "%%%u", addrof->src->id);
+  fprintf(stream, "%%%llu", (unsigned long long)addrof->src->id);
 }
 
 void
@@ -661,7 +663,7 @@ print_instr_call(MirInstrCall *call, FILE *stream)
   if (callee_name)
     fprintf(stream, "@%s", callee_name);
   else
-    fprintf(stream, "%%%u", call->callee->id);
+    fprintf(stream, "%%%llu", (unsigned long long)call->callee->id);
 
   fprintf(stream, "(");
   if (call->args) {
@@ -689,7 +691,7 @@ print_instr_store(MirInstrStore *store, FILE *stream)
   print_instr_head(&store->base, stream, "store");
   assert(store->src && store->src);
   print_comptime_value_or_id(store->src, stream);
-  fprintf(stream, " -> %%%u", store->dest->id);
+  fprintf(stream, " -> %%%llu", (unsigned long long)store->dest->id);
   // print_comptime_value_or_id(store->dest, stream);
 }
 
@@ -711,7 +713,7 @@ print_instr_block(MirInstrBlock *block, FILE *stream)
 #if BL_DEBUG
   fprintf(stream, "%%%s_%u (%u):", block->name, block->base.id, block->base.ref_count);
 #else
-  fprintf(stream, "%%%s_%llu:", block->name, block->base.id);
+  fprintf(stream, "%%%s_%llu:", block->name, (unsigned long long)block->base.id);
 #endif
   if (!block->base.ref_count)
     fprintf(stream, " // NEVER REACHED\n");
@@ -740,7 +742,7 @@ print_instr_fn_proto(MirInstrFnProto *fn_proto, FILE *stream)
   if (fn->llvm_name)
     fprintf(stream, "@%s ", fn->llvm_name);
   else
-    fprintf(stream, "@%u ", fn_proto->base.id);
+    fprintf(stream, "@%llu ", (unsigned long long)fn_proto->base.id);
 
 #if BL_DEBUG
   fprintf(stream, "~%llu (%d) : ", (unsigned long long)fn_proto->base._serial, fn->ref_count);
